requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
altair==5.1.1
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
exceptiongroup==1.1.3
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.7
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.0
pyee==11.0.0
Pygments==2.16.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.12.0
six==1.16.0
sklearn==0.0.post9
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
streamlit==1.26.0
streamlit-webrtc==0.47.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
zipp==3.16.2
